Add ability to run kube tests

+++ src/pfe/file-watcher/server/test/performance-test/run.sh
@@ -29,6 +29,7 @@ CONVERT_ONLY=false
 
 CODEWIND_REPO=git@github.com:eclipse/codewind.git
 RELEASE_DIR="performance-test-codewind"
+USER_DEVFILE=
 
 function usage {
     me=$(basename $0)
@@ -36,13 +37,14 @@ function usage {
 Usage: $me: [-<option letter> <option value> | -h]
 Options:
     --clean-run         To perform a clean run. Default: n
+    --devfile           The devfile to set the workspace on kube with. Required if environment is "kube"
     --environment       The test environment. Default: local
     --post-clean        To perform a post clean up. Default: n 
     --release           The release branch to use. Default: master
     --iterations        Number of iterations to run the performance test. Default: 10
     --convert-only      Convert data to csv only. Will not run performance test and use existing data json on path. Default: false
     --repo              The upstream repo to use. Default: git@github.com:eclipse/codewind.git
-    -h | --help             Display the man page
+    -h | --help         Display the man page
 EOF
 }
 
@@ -84,6 +86,9 @@ while :; do
         --repo=?*)
         CODEWIND_REPO=${1#*=}
         ;;
+        --devfile=?*)
+        USER_DEVFILE=${1#*=}
+        ;;
         -h|--help)
         usage
         exit
@@ -99,6 +104,10 @@ TEST_DIR=$(cd .. && pwd)
 # in local we don't need a clean run from the test as it is done as a pre setup for the performance test
 if [ "$CLEAN_RUN" == "y" ] && [ "$TEST_ENV" == "local" ]; then
     CLEAN_RUN="n"
+fi
+
+if [ "$TEST_ENV" == "kube" ] && [ -z "$USER_DEVFILE" ]; then
+    displayMsg 1 "User devfile is a required parameter on kube". true
 fi
 
 export TURBINE_PERFORMANCE_TEST=${RELEASE_BRANCH}
@@ -120,29 +129,31 @@ if [[ $CONVERT_ONLY == false ]]; then
     do
         echo -e "${CYAN}> Iteration: $run ${RESET}"
 
-        echo -e "${CYAN}> Switching to release directory ${RESET}"
-        cd "$CURR_DIR/$RELEASE_DIR"
-        displayMsg $? "Failed to switch release directory." true
+        if [ "$TEST_ENV" == "local" ]; then
+            echo -e "${CYAN}> Switching to release directory ${RESET}"
+            cd "$CURR_DIR/$RELEASE_DIR"
+            displayMsg $? "Failed to switch release directory." true
 
-        echo -e "${CYAN}> Stopping codewind ${RESET}"
-        ./stop.sh
-        displayMsg $? "Failed to stop codewind." true
+            echo -e "${CYAN}> Stopping codewind ${RESET}"
+            ./stop.sh
+            displayMsg $? "Failed to stop codewind." true
 
-        echo -e "${CYAN}> Running docker system prune ${RESET}"
-        docker system prune -af
-        displayMsg $? "Failed to perform docker system prune." false
+            echo -e "${CYAN}> Running docker system prune ${RESET}"
+            docker system prune -af
+            displayMsg $? "Failed to perform docker system prune." false
 
-        echo -e "${CYAN}> Running docker image prune ${RESET}"
-        docker image prune -af
-        displayMsg $? "Failed to perform docker image prune." false
+            echo -e "${CYAN}> Running docker image prune ${RESET}"
+            docker image prune -af
+            displayMsg $? "Failed to perform docker image prune." false
 
-        echo -e "${CYAN}> Starting codewind ${RESET}"
-        ./run.sh
-        displayMsg $? "Failed to start codewind." true
+            echo -e "${CYAN}> Starting codewind ${RESET}"
+            ./run.sh
+            displayMsg $? "Failed to start codewind." true
+        fi
 
         cd $TEST_DIR
 
-        ./test.sh -t $TEST_ENV -s functional -c $CLEAN_RUN -p $POST_CLEAN
+        NAMESPACE=${NAMESPACE} CLUSTER_IP=${CLUSTER_IP} CLUSTER_PORT=${CLUSTER_PORT} CLUSTER_USER=${CLUSTER_USER} CLUSTER_PASSWORD=${CLUSTER_PASSWORD} ./test.sh -t $TEST_ENV -s functional -c $CLEAN_RUN -p $POST_CLEAN
     done
 fi
